Add support for git repository download; improved parsing; add hugo export
The git import allows to retrieve the last change (timestamp) of each
plugin via git's history.

Multiple improvements of the plugin parsing:
* fix capitalization of headings
* determine a suitable summary for each plugin
* determine the programming language of each plugin

The hugo export is fairly basic for now.  The "hugo-base" directory
needs to be populated with a suitable theme and custom layout files for
providing a good overview over the plugins.